Use u64 instead of usize and/or i64 in TableSummary counts (#32480)
GitOrigin-RevId: 7718141d1b6d6bd5b2198e526a6be82c7d53939a

+++ crates/application/src/snapshot_import/table_change.rs
@@ -6,8 +6,8 @@ use value::TableName;
 
 pub struct TableChange {
     pub added: u64,
-    pub deleted: usize,
-    pub existing: usize,
+    pub deleted: u64,
+    pub existing: u64,
     pub unit: &'static str,
     pub is_missing_id_field: bool,
 }
